+++ static/oils/assets/circulation/onestop/constants/ActionTypes.js
@@ -1,1 +1,9 @@
-export const PATRON_LOOKUP = 'PATRON_LOOKUP';
+export const LOOKUP_PATRON = 'LOOKUP_PATRON';
+export const CHECKOUT_RESOURCE = 'CHECKOUT_RESOURCE';
+export const FAIL_PATRON_LOOKUP = 'FAIL_PATRON_LOOKUP';
+export const FAIL_CHECKOUT = 'FAIL_CHECKOUT';
+export const SUCCESS_CHECKOUT = 'SUCCESS_CHECKOUT';
+export const RETURN_RESOURCE = 'RETURN_RESOURCE';
+export const RENEW_RESOURCE = 'RENEW_RESOURCE';
+export const REQUEST_PATRON = 'REQUEST_PATRON';
+export const RECEIVE_PATRON = 'RECEIVE_PATRON';
