+++ static/oils/assets/circulation/onestop/reducers/index.js
@@ -1,0 +1,12 @@
+import {
+  bindActionCreators,
+  combineReducers,
+} from 'redux'
+
+import patronReducer from './onestop'
+
+const rootReducer = combineReducers({
+  patron: patronReducer,
+});
+
+export default rootReducer;
