+++ static/oils/assets/circulation/onestop/store/configureStore.js
@@ -1,0 +1,19 @@
+import { createStore, applyMiddleware } from 'redux'
+import createLogger from 'redux-logger'
+import thunkMiddleware from 'redux-thunk'
+import rootReducer from '../reducers'
+
+
+const loggerMiddleware = createLogger();
+
+export default function configureStore(initialState) {
+  
+  const store = createStore(
+    rootReducer,
+    applyMiddleware(
+      thunkMiddleware,
+      loggerMiddleware
+    ));
+
+  return store;
+};
